Introduce ConverterHelpers to simplify PresentationConverter
